init service & controller for all service

+++ src/main/java/com/java/service/VetService.java
@@ -9,7 +9,7 @@ public interface VetService {
 
     List<Vet> findAllVet();
 	
-    Vet findVetById(Integer id);
+    Vet findVetById(Long id);
 	
 	Vet addVet(Vet vet);
